+++ src/Event/InsertRepositoryEvent.php
@@ -41,7 +41,7 @@ class InsertRepositoryEvent extends RepositoryEvent
         foreach ($this->entities as $entity) {
             $newRow = $query->insert($entity);
             if ($newRow instanceof ActiveRow) {
-                $entity->internalData($newRow->toArray(), false);
+                $entity->setInternalData($newRow->toArray());
                 $entity->setTable($newRow->getTable());
                 if (count($this->entities) === 1) {
                     return new InsertEventResponse($this, $entity);